Added support for header customization
