+++ packages/notion-utils/src/index.ts
@@ -12,3 +12,4 @@ export * from './id-to-uuid'
 export * from './uuid-to-id'
 export * from './get-all-pages-in-space'
 export * from './get-canonical-page-id'
+export * from './get-page-breadcrumbs'
