+++ packages/react-notion-x/src/renderer.tsx
@@ -36,7 +36,7 @@ export interface NotionRendererProps {
   className?: string
   bodyClassName?: string
 
-  header?: React.ReactNode
+  header?: React.ElementType
   footer?: React.ReactNode
   pageHeader?: React.ReactNode
   pageFooter?: React.ReactNode
@@ -50,7 +50,7 @@ export interface NotionRendererProps {
 interface NotionBlockRendererProps {
   className?: string
   bodyClassName?: string
-  header?: React.ReactNode
+  header?: React.ElementType
   footer?: React.ReactNode
 
   blockId?: string
